move firmware stuff
